change  type latitude and longitude

+++ app/schemas/establishment_schema.py
@@ -7,7 +7,7 @@ class EstablishmentSchema(ma.Schema):
     name = fields.Str(required=True, validate=validate.Length(0, 100),allow_none=False)
     description = fields.Str(required=False, validate=validate.Length(0, 100),allow_none=True)
     about = fields.Str(required=False, validate=validate.Length(0, 300),allow_none=True)
-    latitude = fields.Int(required=True,allow_none=False)
-    longitude = fields.Int(required=True,allow_none=False)
+    latitude = fields.Float(required=True,allow_none=False)
+    longitude = fields.Float(required=True,allow_none=False)
